guacamole js: Stop timer when job ended
In order to display job output to the user as it happens,
we set a timer to refresh the job information every second.

Let's reduce the timer to 0.5 second, and let's clear it
when the test is complete. This way we won't keep calling
the REST API out of necessity when displaying a job.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>